Adding auth functionality and writing tests
